+++ src/polymer/elements/iron-flex-layout-classes.ts
@@ -28,14 +28,16 @@ const styleModules = [
     module: 'iron-flex',
     selector: elementSelectorToPredicate(
         '.layout.horizontal, .layout.vertical, .layout.inline, .layout.wrap,' +
-        '.layout.no-wrap, .layout.center, .layout.center-center, ' +
-        '.layout.center-justified, .flex, .flex-auto, .flex-none')
+            '.layout.no-wrap, .layout.center, .layout.center-center, ' +
+            '.layout.center-justified, .flex, .flex-auto, .flex-none',
+        true)
   },
   {
     module: 'iron-flex-reverse',
     selector: elementSelectorToPredicate(
         '.layout.horizontal-reverse, .layout.vertical-reverse, ' +
-        '.layout.wrap-reverse')
+            '.layout.wrap-reverse',
+        true)
   },
   {
     module: 'iron-flex-alignment',
@@ -43,24 +45,27 @@ const styleModules = [
     // as they're already defined in the `iron-flex` module.
     selector: elementSelectorToPredicate(
         '.layout.start, .layout.end, .layout.baseline, .layout.start-justified, ' +
-        '.layout.end-justified, .layout.around-justified, .layout.justified, ' +
-        '.self-start, .self-center, .self-end, .self-stretch, .self-baseline, ' +
-        '.layout.start-aligned, .layout.end-aligned, .layout.center-aligned, ' +
-        '.layout.between-aligned, .layout.around-aligned')
+            '.layout.end-justified, .layout.around-justified, .layout.justified, ' +
+            '.self-start, .self-center, .self-end, .self-stretch, .self-baseline, ' +
+            '.layout.start-aligned, .layout.end-aligned, .layout.center-aligned, ' +
+            '.layout.between-aligned, .layout.around-aligned',
+        true)
   },
   {
     module: 'iron-flex-factors',
     // Skip `.flex` as it's already defined in the `iron-flex` module.
     selector: elementSelectorToPredicate(
         '.flex-1, .flex-2, .flex-3, .flex-4, .flex-5, .flex-6, .flex-7, ' +
-        '.flex-8, .flex-9, .flex-10, .flex-11, .flex-12')
+            '.flex-8, .flex-9, .flex-10, .flex-11, .flex-12',
+        true)
   },
   {
     module: 'iron-positioning',
     // Skip `[hidden]` as it's a too generic selector.
     selector: elementSelectorToPredicate(
         '.block, .invisible, .relative, .fit, body.fullbleed, ' +
-        '.scroll, .fixed-bottom, .fixed-left, .fixed-top, .fixed-right')
+            '.scroll, .fixed-bottom, .fixed-left, .fixed-top, .fixed-right',
+        true)
   }
 ];
 
@@ -194,17 +199,17 @@ function getMissingStyleModules(
   let missingModules = '';
   for (const [module, nodes] of modules) {
     if (includes.indexOf(module) === -1) {
-      nodes.forEach((node: dom5.Node) => {
-        warnings.push(new Warning({
-          code: 'iron-flex-layout-classes',
-          message: `"${module}" style module is used but not imported.
+      warnings.push(...nodes.map(
+          (node: dom5.Node) => new Warning({
+            code: 'iron-flex-layout-classes',
+            message: `"${module}" style module is used but not imported.
 Import it in the template style include.`,
-          parsedDocument,
-          severity: Severity.WARNING,
-          sourceRange:
-              parsedDocument.sourceRangeForAttributeValue(node, 'class')!
-        }));
-      });
+            parsedDocument,
+            severity: Severity.WARNING,
+            // Prefer warning on class$, as it will override any value of class.
+            sourceRange: parsedDocument.sourceRangeForAttributeValue(
+                node, `class${dom5.hasAttribute(node, 'class$') ? '$' : ''}`)!
+          })));
       missingModules += ' ' + module;
     }
   }